==> 01_millionsong_databasemerge.ipynb <==
### Importing and prepping the dataset

# importando bibliotecas 
import numpy as np
import pandas as pd
#importando 

song_data = pd.read_csv(r'C:\Users\Nathan\Documents\music recommendation system\song_data.csv')
song_data2 = pd.read_fwf(r'C:\Users\Nathan\Documents\music recommendation system\10000.txt')
song_data2.columns = ['user_id','song_id','listen_count']

song_df = pd.merge(song_data2, song_data.drop_duplicates(['song_id']), on="song_id", how="left")

song_df.to_csv(r'C:\Users\Nathan\Documents\music recommendation system\song_df.csv', index = False)
