==> 03_Correcting_genres_dataset.ipynb <==
## Correcting Music Genre Dataset
Due to multiple connection errors during the webscrapping it was necessary to divided the webscrapping in parts in order to avoid the lost of time and computer processing.
For now lets append those separated files.


import pandas as pd
import numpy as np

all_genres0 = pd.read_excel(r"C:\Users\Nathan\Documents\music recommendation system\allgenres1.xlsx",header=None)
all_genres0 = all_genres0[1]
all_genres1 = pd.read_excel(r'C:\Users\Nathan\Documents\music recommendation system\allgenres2.xlsx',header=None)
all_genres1 = all_genres1[1]
all_genres1 = all_genres1[1:]
all_genres2 = pd.read_excel(r'C:\Users\Nathan\Documents\music recommendation system\allgenres3.xlsx',header=None)
all_genres2 = all_genres2[1]
all_genres2 = all_genres2[1:]
all_genres = all_genres0.append(all_genres1.append(all_genres2))
all_genres.reset_index(drop=True)

all_genres0[0:10]

all_genres.to_excel(r'C:\Users\Nathan\Documents\music recommendation system\allgenres.xlsx',header=None)
