==> 05_Dashboard_Data_Preparation.ipynb <==
import pandas as pd
import numpy as np
#allgenres = pd.read_excel(r'C:\Users\Nathan\Documents\music recommendation system\allgenres.xlsx',header=None)
song_df = pd.read_csv(r'C:\Users\Nathan\Documents\music recommendation system\songs_subset.csv',header=None)
#genres1 = pd.read_excel(r'C:\Users\Nathan\Documents\music recommendation system\clean_genres.xlsx',header=None)
#genres2 = pd.read_excel(r'C:\Users\Nathan\Documents\music recommendation system\clean_genres2.xlsx',header=None)

### Grouping the genres obtained by the artists on the original dataset

df_genres = [] 
df_genres = [genres[j] for i in range(0,len(song_df)) for j in range(0,len(artists)) if song_df['artist_name'][i] == artists[j]]  

song_df.rename(columns={1: "user_id",2:"song_id",3:"listen_count",4:"title",5:"release",6:"artist_name",7:"year",8:"genre"},inplace=True)
song_df["song"] = song_df["title"] + " - " + song_df["artist_name"]

### Cleaning the songs without year or genre

song_df = song_df[song_df['year']!=0]

song_df = song_df.dropna()

song_df['genre']

genres = [None]*len(song_df)

song_df.reset_index(drop=True,inplace=True)

### Making List to String

import re
for i in range(0,len(song_df['genre'])):
        genre = ''
        str1 = re.findall("'(.*?)'",song_df['genre'][i])
        for j in range(0,len(str1)):
            genre = genre + '|' + str1[j]
        genre = genre[1:]
        genres[i] = genre

song_df['genre'] = genres

### grouping most popular songs

song_grouped = song_df.groupby(['song']).agg({'listen_count': 'count'}).reset_index()
top_songs = song_grouped.sort_values(by=["listen_count"],ascending=False)
top_songs.reset_index(drop=True,inplace=True)

top_songs

### df de song_df sem musicas repetidas
reduced_df = song_df.drop_duplicates(subset='song')
reduced_df.reset_index(drop=True,inplace=True)

###incluindo genero e ano em topsongs
topsongs_genres = []
topsongs_year = []
for i in range(0,len(top_songs)):
    for j in range(0,len(reduced_df)):
        if top_songs['song'][i] == reduced_df['song'][j]:
            topsongs_genres.append(reduced_df['genre'][j])
            topsongs_year.append(reduced_df['year'][j])

top_songs['genre'] = topsongs_genres
top_songs['year'] = topsongs_year

top_songs.to_excel(r'C:\Users\Nathan\Documents\music recommendation system\top_songs_complete.xlsx')

teste = pd.read_excel(r'C:\Users\Nathan\Documents\music recommendation system\top_songs_complete.xlsx')

top_songs['genre'][0]

### Checking which is the most popular genres

all_genres = []

for i in range(0,len(top_songs)):
    if type(re.search('|',genres[i])) == re.Match:
        str1 = re.findall('(.*?)\|',top_songs['genre'][i])
        str2 = re.findall('\|([^|]*)$',top_songs['genre'][i])

        for j in range(0,len(str1)):
            all_genres.append(str1[j])
        if len(str2) == 1:
            all_genres.append(str2[0])

    else:
        all_genres.append(genres[i])

all_genres = pd.Series(all_genres)

top_genres = all_genres.value_counts()

top_25_genres = top_genres[0:25]

top_25_genres

type(top_25_genres.index)

top25list = top_25_genres.index.tolist()

### Use only the 25 most popular genre, otherwise it will be categorized as 'other'


def check_top_25(string,top_25):
    check = 0
    for i in range(0,25):        
        if string == top_25[i]:
            check = 1
        
    if check == 0:
        string = 'other'
    
    return string
    
        

genres = [None]*len(top_songs)
for i in range(0,len(top_songs)):
    if type(re.search('|',top_songs['genre'][i])) == re.Match:
        str1 = re.findall('(.*?)\|',top_songs['genre'][i])
        for j in range(0,len(str1)):
            str1[j] = check_top_25(str1[j],top25list)  

        str2 = re.findall('\|([^|]*)$',top_songs['genre'][i])
        str2 = check_top_25(str2,top25list)
        str2 = [str2]

        
        genres[i] = str1 + str2
        genres[i] = list(dict.fromkeys(genres[i]))
    else:
        genres[i] = check_top_25(top_songs['genre'][i],top25list)

top_songs['genre'] = genres

top_songs = top_songs.explode('genre')

top_songs.to_excel(r'C:\Users\Nathan\Documents\music recommendation system\dashboard_topsongs.xlsx')
